pyproject.toml
[project]
name = "credit-simulator"
version = "0.1.0"
description = ""
authors = [
    {name = "eden",email = "yangsehyuk1117@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=3.0.2,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
